requirements.txt
pygame-ce
sqlite3
sqlalchemy
matplotlib
seaborn
pandas
numpy